Refactor HashAddress into HashStore module, update pytests and update poetry dependencies
