+++ src/hashstore/__init__.py
@@ -15,8 +15,7 @@ processes that function across a cluster environment. Some properties:
     their persistent identifier (PID)
 """
 
-from hashstore.hashstore import HashStore
-from hashstore.hashaddress import HashAddress
+from hashstore.hashstore import HashStore, HashAddress
 from hashstore.hashstore_factory import HashStoreFactory
 
 __all__ = ("HashStore", "HashAddress", "HashStoreFactory")
